requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
django-ckeditor==5.5.0
django-cors-headers==2.2.0
django-crontab==0.7.1
django-filter==1.1.0
django-haystack==2.8.1
django-js-asset==1.1.0
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-haystack==1.8.3
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==2.7
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
mutagen==1.40.0
Pillow==5.1.0
PyJWT==1.6.4
PyMySQL==0.8.1
python-dateutil==2.7.3
pytz==2018.4
redis==2.10.6
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
